+++ src/infra/adapters/appollo-server/type.ts
@@ -1,7 +1,0 @@
-import { NonEmptyArray } from 'type-graphql';
-import { Container } from '../../../main/container';
-
-export type ResolverType = {
-  resolver: NonEmptyArray<Function>;
-  container: Container
-};