+++ src/infra/adapters/index.ts
@@ -5,4 +5,3 @@ export * from './kenx/interface';
 export * from './kenx/knex';
 export * from './redis/redis';
 export * from './appollo-server/appolo-server';
-export * from './appollo-server/type';